+++ src/main/java/car/Car.java
@@ -1,4 +1,4 @@
-package model;
+package car;
 
 public class Car {
 }
