+++ src/main/java/car/CarCreator.java
@@ -1,4 +1,4 @@
-package model;
+package car;
 
 import java.util.*;
