+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,8 +1,5 @@
 package racingcar;
 
-import model.RacingCarGame;
-import view.InputView;
-
 public class RacingCarApplication {
 
     public static void main(String[] args) {
